Refactored twin to clean code
